backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.1",
    "openai>=2.6.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
    "slowapi>=0.1.9",
    "psycopg2-binary>=2.9.9",
    "SQLAlchemy>=2.0.25",
    "alembic>=1.13.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
]
